Use np.lib.format.open_memmap rather than shared memory
Doing it this way doesn't require keeping a 1.6GB file loaded in RAM in
a separate process.  Some quick benchmarks showed it to be about as fast
as shared memory on my laptop once the data was already loaded into the
disk cache; this may be different on a server where it might get evicted
from the cache if it's accessed infrequently.